# perf(metainfo): avoid temp maps to optimize SaveMetaInfoToMap